Break transects module into top-level and viz parts
